requirements.txt
Flask
sqlalchemy
pyyaml
